pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "brave-torch"
version = "4.7.9"
description = "Swarms - Pytorch"
license = "MIT"
authors = ["Kye Gomez <kye@apac.ai>"]
homepage = "https://github.com/kyegomez/BRAVE-ViT-Swarm"
documentation = "https://swarms.apac.ai"  
readme = "README.md" 
repository = "https://github.com/kyegomez/BRAVE-ViT-Swarm"
keywords = ["artificial intelligence", "deep learning", "optimizers", "Prompt Engineering", "swarms", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10"
]


[tool.poetry.dependencies]
python = ">=3.9,<4.0"
torch = ">=2.1.1,<3.0"
zetascale = "*"
einops = "*"
einops_exts = "*"

[tool.poetry.dev-dependencies]
black = "23.3.0"

[tool.poetry.group.lint.dependencies]
ruff = ">=0.0.249,<0.3.5"
types-toml = "^0.10.8.1"
types-pytz = "^2023.3.0.0"
black = "^23.1.0"
types-chardet = "^5.0.4.6"
mypy-protobuf = "^3.0.0"


[tool.ruff]
line-length = 70
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.black]
line-length = 70
target-version = ['py38']
preview = true